Adaptação para utilizar o mapa
-> PRINCIPAL: adaptei o código para utilizá-lo na geração do mapa com a mathplotlib;

-> Renomeei o código " vizinho_prox_firstRandom " para " vizinho_prox_first_random " com o intuito de seguir o parâmetro snake_case;

-> Ajustei o código sob as orientações da biblioteca "Pylint";
